Limit how many threads can prepare a statement
`StatementCache` aims to prevent redundant SQL statement preparation
with a simple cache. However in cases where many, many writers are
racing to prepare a statement, lots of redundant and expensive work can
be done because `tokio_postgres::client::Client::prepare_typed` is
called before `StatementCache::insert`, and only the latter takes the
lock on the cache.

This commit allows entries in the `StatementCache` map to be either
empty, a Semaphore, or a prepared Statement. The first call to
`prepare_typed()` will insert a `Semaphore`. Then, tasks will wait for a
semaphore permit before preparing the query, to ensure that only one
task sends a `PREPARE` statement at once. At both steps in the process,
the map will be read via a read lock first, then again via a write lock,
before updating the entry.